build(lint): type-aware linting, and a rule-by-rule reading of what was off
`.oxlintrc.json` enabled `correctness` and made a measured case for leaving the
wider categories off. That case is still sound as a *category* argument, and it
was only ever a category argument: no rule outside `correctness` had been
considered on its own. Twenty-one now have been, each measured, each on or off
with the reason written into the config so nobody re-derives it.

`oxlint-tsgolint` is installed, and `options.typeAware` turns on the rules that
can ask what a value *is* rather than what it looks like. That is worth more here
than in most codebases: the difficulty in this one is values that are
structurally similar and semantically opposite.

`test/tsconfig.json` is new and load bearing. tsgolint resolves a tsconfig by
walking up from each file; from `test/` that found the source project, which
excludes `test/`, so the whole tree was linted on default compiler options.
Measured: `no-unnecessary-condition` reported 71 findings before that file and 24
after, the difference being every `noUncheckedIndexedAccess` guard in a test read
as a condition that can never fire.

Turning `typeAware` on surfaced four `correctness` errors that had never run:

- `render/paint.ts` declared a parameter as `LineEnd['width'] | LineEnd['length']`,
  which is one union spelled twice. OOXML gives `@w` and `@len` the same three
  classes, and the comment now says so.
- `parse/edits.ts` spread a string to count its code points. That is the right
  reading, since a glyph outside the BMP is one character that `String.length`
  calls two, and `Array.from` iterates identically without reading as the
  accidental spread the rule is about.
- `heuristic/frame.ts` tested `doc.fonts?.ready` for truthiness. `doc.fonts` is
  what may be absent; `ready` is a promise, and a promise is always truthy.
- the remaining two are `String(value || '')` at the heuristic lane's two host
  boundaries, where `[object Object]` for an object is the intended answer. Off
  in that directory's override, with the reason.

On, and clean across the repo: `no-floating-promises`, `no-misused-promises`,
`await-thenable`, `switch-exhaustiveness-check`, `restrict-template-expressions`.
Zero findings each, which is the point: every entry point in the loop is async,
`AssetResolver` returns a union whose whole risk is a dropped `await`, and every
SVG attribute in `src/render/` is built by template literal, where an `undefined`
emits the word into the document while the island beside it stays perfect.

The `no-unsafe-*` family measured 578 findings, and all of them land in
`src/heuristic/`, in untyped JS (`examples/`, `charcheck.config.js`), or at a
`JSON.parse` config boundary in `site/scripts/`. Scoped to `src/**` and switched
off again for the heuristic lane, they become a gate rather than a backlog: the
loop is at zero today, so the next `any` to reach it fails the build.

`no-unnecessary-type-assertion` found 13 and all 13 are gone. Five were in the
loop, including a `structuredClone(deck) as DeckIr` that TypeScript 7 types
generically, and four `svg as unknown as Element` double casts on a value that is
already an `Element`.

`prefer-structured-clone` found 12. Two were the loop cloning a `RenderIr` by
JSON round trip, and those now go through `cloneIr`, a new export beside `emuOf`
and `inchesOf`. The spellings are not equivalent in general, because
`JSON.parse(JSON.stringify(...))` drops `undefined`-valued keys and cannot carry
a `Uint8Array`; they agree on this model only because of its JSON contract, which
makes the difference invisible today and a silent divergence the moment that
contract slips. Three of the twelve are sites where the JSON round trip *is* the
assertion; those keep it, with a scoped disable saying why, because switching
them would leave a test that passes while proving something else.

`no-array-sort` found 8, every one on a freshly built array, so `toSorted` reads
the same and states the intent.

Five rules the plan expected to enable are off instead, because measuring them
changed the answer, and the config says so at each:

- `require-await` (8). Dropping `async` is not behaviour-preserving: an `async`
  function turns a synchronous throw in its body into a rejected promise, and
  `convertDeck(...).catch(...)` is how a consumer would expect to see one.
- `no-useless-undefined` (2). Both are load bearing, and the suggested fix at
  `bytesFor: () => undefined` infers `void` and does not typecheck.
- `consistent-function-scoping` (28). 23 inside the stringified extractor
  payload; the rest are test helpers scoped to the `describe` that reads them.
- `no-unnecessary-condition` (24). It reports `element.textContent ?? ''` as
  dead, and `Node.textContent` is `string | null` in the lib `tsc` reads. A gate
  that calls a necessary guard unnecessary teaches people to delete correct code.
- `default-case` (8), unchanged from the plan's reading and worth restating: all
  eight are exhaustive switches over closed unions whose *absent* default is what
  makes adding a member a compile error.

`lefthook.yml` gains the cost. oxlint now builds the program before it lints, so
the pre-commit job is about eight seconds whether one file is staged or fifty.
Verified rather than assumed: a single-file `--fix` neither fails nor hangs.

+++ test/oracle/writer-output.test.ts
@@ -43,7 +43,7 @@ async function slidesOf(bytes: Uint8Array): Promise<Map<string, string>> {
 	const slides = new Map<string, string>()
 	for (const [name, data] of entries) {
 		if (!/^ppt\/slides\/slide\d+\.xml$/i.test(name)) continue
-		slides.set(name, typeof data === 'string' ? data : decoder.decode(data as Uint8Array))
+		slides.set(name, typeof data === 'string' ? data : decoder.decode(data))
 	}
 	return slides
 }
